Simplify Vec2D

+++ src/animation/state_machine_instance.cpp
@@ -240,11 +240,11 @@ void StateMachineInstance::processEvent(Vec2D position, EventType hitEvent) {
     position -= Vec2D(m_ArtboardInstance->originX() * m_ArtboardInstance->width(),
                       m_ArtboardInstance->originY() * m_ArtboardInstance->height());
 
-    const int hitRadius = 2;
-    auto hitArea = AABB(position.x() - hitRadius,
-                        position.y() - hitRadius,
-                        position.x() + hitRadius,
-                        position.y() + hitRadius)
+    const float hitRadius = 2;
+    auto hitArea = AABB(position.x - hitRadius,
+                        position.y - hitRadius,
+                        position.x + hitRadius,
+                        position.y + hitRadius)
                        .round();
 
     for (const auto& hitShape : m_HitShapes) {
